backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
pinecone>=5.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
pymupdf>=1.24.0
fastapi-azure-auth>=5.0.0
tenacity>=8.0.0
PyJWT>=2.0.0
